+++ scaled_heatmaps.py
@@ -85,7 +85,7 @@ def display_clustermap(sort_via: str, primary_only: bool = False, separate_bioma
     exposure = val_df[sort_via]                                         # extract pd.Series of exposure metrics
 
     df_plot = val_df.transpose()                                        # transpose our dataframe: a column is now a patient
-    df_plot.columns = df_plot.iloc[0]                                   # set our column labels to be our patient IDs
+    df_plot.columns = df_plot.iloc[0].astype(int)                       # set our column labels to be our patient IDs
     df_plot = df_plot.iloc[2:]                                          # get rid of the patient IDs row and exposure metric rows
 
     # create objects for coloring patient exposure, p-values, and rho values
@@ -234,7 +234,7 @@ def save_clustermap(sort_via: str, primary_only: bool = False, separate_biomarke
     exposure = val_df[sort_via]                                         # extract pd.Series of exposure metrics
 
     df_plot = val_df.transpose()                                        # transpose our dataframe: a column is now a patient
-    df_plot.columns = df_plot.iloc[0]                                   # set our column labels to be our patient IDs
+    df_plot.columns = df_plot.iloc[0].astype(int)                       # set our column labels to be our patient IDs
     df_plot = df_plot.iloc[2:]                                          # get rid of the patient IDs row and exposure metric rows
 
     # create objects for coloring patient exposure, p-values, and rho values
